Refactored TestRunner api, now with a basic class because constructors cannot be modelled with interfaces directly
