--- src/api/test-runner/Result.ts
@@ -1,8 +1,0 @@
-enum Result{
-  SUCCESS,
-  FAILED,
-  TIMED_OUT,
-  ERROR
-}
-
-export default Result;
