--- src/api/test-runner/TestResultMap.ts
@@ -1,7 +1,0 @@
-import TestResult from './TestResult';
-
-interface TestResultMap{
-  [id: string]: TestResult;
-}
-
-export default TestResultMap;
